Clean up tests: move TestTaskTool to separate file
